requirements.txt
beautifulsoup4==4.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
fake-useragent==1.1.3
html5lib==1.1
idna==3.4
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
webencodings==0.5.1
